+++ src/pages/Basket/BasketTotal/BasketTotal.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from 'react-i18next';
 import styles from './BasketTotal.module.scss';
 import Button from '../../../components/Button/Button';
 
-function BasketTotal({ price }: { price: number }) {
+function BasketTotal({ price, onPayOrderClick }: { price: number, onPayOrderClick: () => void }) {
     const { t } = useTranslation();
     return (
         <div className={styles.basket_total}>
@@ -11,7 +11,10 @@ function BasketTotal({ price }: { price: number }) {
                 <span className={styles.basket_total__price}>{`${price} ₸`}</span>
             </div>
             <div className={styles.basket_total__button}>
-                <Button>{t(`pages.basket.pay`)}</Button>
+                <Button
+                    onClick={onPayOrderClick}>
+                    {t(`pages.basket.pay`)}
+                </Button>
             </div>
         </div>
     );
