+++ src/pages/WaitingOrderPage/OrderListArticle/OrderListArticle.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .orderListArticle {
     display: flex;
