+++ src/pages/WaitingOrderPage/OrderListArticle/OrderListArticle.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
+import { OrderState } from '../../../utils/api/orderService/orderService';
 import { OrderListItem } from '../OrderListItem/OrderListItem';
 import styles from './OrderListArticle.module.scss';
-import { OrderState } from '../../utils/api/orderService/orderService';
 
 type OrderListArticleProps = {
     order: OrderState;
