+++ src/pages/WaitingOrderPage/OrderListItem/OrderListItem.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .orderListItem {
     &__item {
