+++ src/pages/WaitingOrderPage/WaitingOrder/WaitingOrder.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .waitingOrder {
     &__title {
