+++ src/pages/WaitingOrderPage/WaitingOrderPage.tsx
@@ -1,5 +1,5 @@
-import WaitingOrder from '../WaitingOrder/WaitingOrder';
 import { OrderProvider } from '../../contexts/OrderContext';
+import WaitingOrder from './WaitingOrder/WaitingOrder';
 
 const WaitingOrderPage = () => {
     return (
